MAKE abstract Renderer that provides an interface for creating view-export renderers
